+++ apps/frontend/src/libs/components/sidebar/sidebar.tsx
@@ -1,7 +1,12 @@
 import { Button } from "~/libs/components/components.js";
 import { SIDEBAR_ITEMS } from "~/libs/constants/constants.js";
 import { getValidClassNames } from "~/libs/helpers/helpers.js";
-import { useLocation } from "~/libs/hooks/hooks.js";
+import {
+	useAppDispatch,
+	useCallback,
+	useLocation,
+} from "~/libs/hooks/hooks.js";
+import { actions as authActions } from "~/modules/auth/auth.js";
 
 import { SidebarLink } from "./libs/components/components.js";
 import styles from "./styles.module.css";
@@ -16,6 +21,13 @@ const Sidebar: React.FC<Properties> = ({
 	onSidebarToggle,
 }: Properties) => {
 	const { pathname } = useLocation();
+
+	const dispatch = useAppDispatch();
+
+	const handleSignOut = useCallback(
+		() => void dispatch(authActions.logOut()),
+		[dispatch],
+	);
 
 	return (
 		<div
@@ -37,19 +49,22 @@ const Sidebar: React.FC<Properties> = ({
 
 			<div className={styles["logo-container"]}>Logo</div>
 			<div className={styles["buttons-container"]}>
-				{SIDEBAR_ITEMS.map(({ href, icon, label }) => {
-					const { active, inactive } = icon;
+				<div className={styles["navlinks-container"]}>
+					{SIDEBAR_ITEMS.map(({ href, icon, label }) => {
+						const { active, inactive } = icon;
 
-					return (
-						<SidebarLink
-							iconName={href === pathname ? active : inactive}
-							key={label}
-							label={label}
-							pathname={pathname}
-							to={href}
-						/>
-					);
-				})}
+						return (
+							<SidebarLink
+								iconName={href === pathname ? active : inactive}
+								key={label}
+								label={label}
+								pathname={pathname}
+								to={href}
+							/>
+						);
+					})}
+				</div>
+				<Button label="Sign out" onClick={handleSignOut} />
 			</div>
 		</div>
 	);