+++ apps/frontend/src/libs/types/icon-name.type.ts
@@ -4,6 +4,7 @@ type IconName =
 	| "crossedEye"
 	| "eye"
 	| "menu"
+	| "signOut"
 	| "wheelActive"
 	| "wheelInactive";