+++ apps/frontend/src/libs/types/query-parameters.type.ts
@@ -1,0 +1,5 @@
+type QueryParameters = {
+	[key: string]: string;
+};
+
+export { type QueryParameters };