+++ apps/frontend/src/libs/types/types.ts
@@ -1,13 +1,14 @@
 export { type AsyncThunkConfig } from "./async-thunk-config.type.js";
 export {
+	type Control,
 	type FieldErrors,
 	type FieldPath,
 	type FieldValues,
 	type FormFieldProperties,
 } from "./form-field-properties.type.js";
 export { type IconName } from "./icon-name.type.js";
+export { type QueryParameters } from "./query-parameters.type.js";
 export { type SidebarItem } from "./sidebar-item.type.js";
-export { type Control } from "react-hook-form";
 export {
 	type OnboardingQuestionResponseDto,
 	type ServerErrorDetail,