+++ apps/frontend/src/pages/auth/libs/components/reset-password-form/libs/enums/enums.ts
@@ -1,0 +1,1 @@
+export { ConfirmPasswordCustomValidation } from "shared";