+++ apps/frontend/src/pages/auth/libs/maps/auth-route-to-header.map.ts
@@ -1,0 +1,13 @@
+import { AppRoute } from "~/libs/enums/enums.js";
+import { type ValueOf } from "~/libs/types/types.js";
+
+import { type RoutesWithHeader } from "../types/types.js";
+
+const authRouteToHeader: Record<ValueOf<RoutesWithHeader>, string> = {
+	[AppRoute.FORGOT_PASSWORD]: "forgot password",
+	[AppRoute.RESET_PASSWORD]: "reset password",
+	[AppRoute.SIGN_IN]: "sign in",
+	[AppRoute.SIGN_UP]: "sign up",
+};
+
+export { authRouteToHeader };